src/MOTChallengeEvalKit/MOTS/requirements.txt
cycler==0.10.0
Cython==0.29.20
kiwisolver==1.2.0
matplotlib==3.2.2
numpy==1.19.0
opencv-python==4.2.0.34
pandas==1.0.5
Pillow==7.1.2
pycocotools==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
scipy==1.5.0
six==1.15.0
